Mesaj controllerları revize edildi.

+++ BusinessLayer/Concrete/MessageManager.cs
@@ -39,36 +39,51 @@ namespace BusinessLayer.Concrete
             return values;
         }
 
-        public async Task TAddAsync(Message t)
+        public async Task AddMessageAsync(Message message, string senderUserName, string receiverUserName)
         {
-            t.Details = await TextFileManager.TextFileAddAsync(t.Details, TextFileManager.GetMessageContentFileLocation());
-            t.MessageDate = DateTime.Now;
-            t.MessageStatus = true;
-            await _messageDal.InsertAsync(t);
+            if (senderUserName == receiverUserName)
+                return;
+            var senderUser = await _userService.FindByUserNameAsync(senderUserName);
+            if (senderUser == null)
+                return;
+            message.SenderUserId = senderUser.Id;
+            var receiverUser = await _userService.FindByUserNameAsync(receiverUserName);
+            if (receiverUser == null)
+                return;           
+            message.ReceiverUserId = receiverUser.Id;
+            message.Details = await TextFileManager.TextFileAddAsync(message.Details, TextFileManager.GetMessageContentFileLocation());
+            message.MessageDate = DateTime.Now;
+            message.MessageStatus = true;
+            await _messageDal.InsertAsync(message);
         }
 
-        public async Task TDeleteAsync(Message t)
+        public async Task DeleteMessageAsync(int id, string userName)
         {
-            DeleteFileManager.DeleteFile(t.Details);
-            await _messageDal.DeleteAsync(t);
+            var message = await GetByIdAsync(id);
+            DeleteFileManager.DeleteFile(message.Details);
+            await _messageDal.DeleteAsync(message);
         }
 
-        public async Task<Message> TGetByFilterAsync(Expression<Func<Message, bool>> filter = null)
+        public async Task<Message> GetByFilterAsync(Expression<Func<Message, bool>> filter = null)
         {
             var value = await _messageDal.GetByFilterAsync(filter);
             value.Details = await TextFileManager.ReadTextFileAsync(value.Details);
             return value;
         }
 
-        public async Task<Message> TGetByIDAsync(int id)
+        public async Task<Message> GetByIdAsync(int id)
         {
             var value = await _messageDal.GetByIDAsync(id);
             value.Details = await TextFileManager.ReadTextFileAsync(value.Details);
             return value;
         }
 
-        public async Task TUpdateAsync(Message t)
+        public async Task UpdateMessageAsync(Message t, string userName)
         {
+            var user = await _userService.FindByUserNameAsync(userName);
+            if (user != null)
+                return;
+            t.SenderUserId = user.Id;
             DeleteFileManager.DeleteFile(t.Details);
             t.Details = await TextFileManager.TextFileAddAsync(t.Details, TextFileManager.GetMessageContentFileLocation());
             await _messageDal.UpdateAsync(t);
@@ -118,7 +133,7 @@ namespace BusinessLayer.Concrete
         public async Task<Message> GetSendMessageAsync(int id, Expression<Func<Message, bool>> filter = null)
         {
             var value = await _messageDal.GetSendedMessageAsync(id, filter);
-            if (value!=null)
+            if (value != null)
                 value.Details = await TextFileManager.ReadTextFileAsync(value.Details);
             return value;
         }
